chore: move options to dedicated files
